+++ View/HistoryPage.xaml
@@ -1,14 +1,89 @@
 ﻿<Page x:Class="PinusPengger.View.HistoryPage"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:PinusPengger.View"
+      xmlns:controller="clr-namespace:PinusPengger.ViewModel"
+      xmlns:model="clr-namespace:PinusPengger.Model"
+      xmlns:sys="clr-namespace:System;assembly=mscorlib"
+      xmlns:converter="clr-namespace:PinusPengger.ViewModel.Converter"
       mc:Ignorable="d" 
-      d:DesignHeight="450" d:DesignWidth="800"
-      Title="HistoryPage">
-
-    <Grid>
-        
+      d:DesignHeight="420" 
+      d:DesignWidth="920">
+    <Page.Resources>
+        <controller:Ticker x:Key="ticker"/>
+    </Page.Resources>
+    <Grid Background="White">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="20"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="290"/>
+            <RowDefinition Height="10"/>
+        </Grid.RowDefinitions>
+        <Border Grid.Row="1">
+            <Grid>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="*" />
+                </Grid.ColumnDefinitions>
+                <TextBlock x:Name="jamSkrg"
+                       Grid.Column="0"
+                       HorizontalAlignment="Center"
+                       VerticalAlignment="Center"
+                       Text="{Binding Source={StaticResource ticker},Path=Now, Mode=OneWay, StringFormat='dd MMMM yyyy  hh:mm:ss' }"/>
+            </Grid>
+        </Border>
+        <StackPanel Grid.Row="2"
+                    Margin="5,0,0,0">
+            <TextBlock Text="Masukkan Kode:"
+                       FontWeight="Bold"/>
+            <StackPanel Height="30"
+                        Orientation="Horizontal">
+                <TextBox Width="200"
+                         Height="25"
+                         FontSize="14"
+                         Foreground="DarkGray"
+                         Text="Kode"/>
+                <Button Height="25"
+                        Width="40">
+                    <Path Data="{StaticResource MagnifyingGlassIcon}"
+                          Grid.Column="1"
+                          Stretch="Uniform"
+                          Fill="Black"
+                          Width="15"
+                          Height="15"
+                          HorizontalAlignment="Center"
+                          VerticalAlignment="Center"
+                          Margin="0,0,0,1" />
+                </Button>
+            </StackPanel>
+        </StackPanel>
+        <DataGrid x:Name="tabelHistori"
+                  Margin="10"
+                  Height="290"
+                  AutoGenerateColumns="False"
+                  Grid.Row="3"
+                  CanUserAddRows="True">
+            <DataGrid.Resources>
+                <converter:PersonConverter x:Key="PersonConverter"/>
+                <converter:RoomConverter x:Key="RoomConverter"/>
+            </DataGrid.Resources>
+            <DataGrid.Columns>
+                <DataGridTextColumn Header="Kode Pemesanan" 
+                                    Binding="{Binding KodePemesanan}"/>
+                <DataGridTextColumn Header="Nama" Width="200" 
+                                    Binding="{Binding Person,Converter={StaticResource PersonConverter}}"/>
+                <DataGridTextColumn Header="Nomor Kamar" Width="*"  
+                                    Binding="{Binding Room, Converter={StaticResource RoomConverter}}"/>
+                <DataGridTextColumn Header="Tanggal Checkin" Width="*" 
+                                    Binding="{Binding CheckIn, StringFormat='dd/MM/yyyy'}"/>
+                <DataGridTextColumn Header="Tanggal Checkout" Width="*"  
+                                    Binding="{Binding CheckOut, StringFormat='dd/MM/yyyy'}"/>
+            </DataGrid.Columns>
+        </DataGrid>
     </Grid>
 </Page>
