+++ View/HomeView.xaml
@@ -1,12 +1,12 @@
-﻿<UserControl x:Class="PinusPengger.Controller.HomeController"
+﻿<UserControl x:Class="PinusPengger.ViewModel.HomeController"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:controller="clr-namespace:PinusPengger.Controller"
+             xmlns:controller="clr-namespace:PinusPengger.ViewModel"
              xmlns:model="clr-namespace:PinusPengger.Model"
              xmlns:sys="clr-namespace:System;assembly=mscorlib"
-             xmlns:converter="clr-namespace:PinusPengger.Controller.Converter"
+             xmlns:converter="clr-namespace:PinusPengger.ViewModel.Converter"
              mc:Ignorable="d" 
              d:DesignHeight="420" d:DesignWidth="920">
     <UserControl.Resources>
