+++ View/LoginWindow.xaml
@@ -1,10 +1,10 @@
-﻿<Window x:Class="PinusPengger.Controller.LoginWindowController"
+﻿<Window x:Class="PinusPengger.ViewModel.LoginWindowController"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:view="clr-namespace:PinusPengger.View"
-        xmlns:controller="clr-namespace:PinusPengger.Controller"
+        xmlns:controller="clr-namespace:PinusPengger.ViewModel"
         mc:Ignorable="d"
         Title="Hotel Jaya Abadi - Login Page" 
         Height="540"
