+++ View/MainWindow.xaml
@@ -1,11 +1,11 @@
-﻿<Window x:Class="PinusPengger.Controller.MainWindowController"
+﻿<Window x:Class="PinusPengger.ViewModel.MainWindowController"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:view="clr-namespace:PinusPengger.View" 
         xmlns:model="clr-namespace:PinusPengger.Model"
-        xmlns:controller="clr-namespace:PinusPengger.Controller"
+        xmlns:controller="clr-namespace:PinusPengger.ViewModel"
         mc:Ignorable="d"
         Title="Hotel Jaya Abadi" 
         Height="540" 
@@ -13,7 +13,7 @@
         ResizeMode="CanMinimize"
         WindowStyle="None">
     <Window.DataContext>
-        <controller:CoreController/>
+        <controller:CoreViewModel/>
     </Window.DataContext>
     <Grid>
         <Grid.RowDefinitions>
