+++ View/ReservasiView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="PinusPengger.Controller.ReservasiController"
+﻿<UserControl x:Class="PinusPengger.ViewModel.ReservasiController"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
