+++ ViewModel/Converter/ModelConverter.cs
@@ -7,15 +7,22 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Data;
 
-namespace PinusPengger.Controller.Converter
+namespace PinusPengger.ViewModel.Converter
 {
-    internal class PersonConverter : IValueConverter
+    internal class RoomConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (value is Person person)
+            if (value is Room room)
             {
-                return person.Name;
+                char typeRoom = room.Type switch
+                {
+                    RoomType.Reguler => 'R',
+                    RoomType.VIP => 'V',
+                    RoomType.VVIP => 'P',
+                    _ => 'R',
+                };
+                return $"{typeRoom}.{room.RoomNumber.Floor}.{room.RoomNumber.RoomNumber}";
             }
             return string.Empty;
         }
