+++ ViewModel/Converter/PersonConverter.cs
@@ -7,22 +7,15 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Data;
 
-namespace PinusPengger.Controller.Converter
+namespace PinusPengger.ViewModel.Converter
 {
-    internal class RoomConverter : IValueConverter
+    internal class PersonConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (value is Room room)
+            if (value is Person person)
             {
-                char typeRoom = room.Type switch
-                {
-                    RoomType.Reguler => 'R',
-                    RoomType.VIP => 'V',
-                    RoomType.VVIP => 'P',
-                    _ => 'R',
-                };
-                return $"{typeRoom}.{room.RoomNumber.Floor}.{room.RoomNumber.RoomNumber}";
+                return person.Name;
             }
             return string.Empty;
         }
