+++ ViewModel/MainWindowController.cs
@@ -8,7 +8,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 
-namespace PinusPengger.Controller
+namespace PinusPengger.ViewModel
 {
     /// <summary>
     /// Interaction logic for MainWindow.xaml
